fix insights aspect to have the modern aspect foramt
